requirements.txt
biopython
chromadb
sentence-transformers
openai
python-dotenv
tenacity
pydantic
numpy